SDU: Remove dependency to null termination.
Previous implementation was able to send only null terminated
strings. The new implementation uses the first byte of the circular
buffer to hold the length of the buffer and uses
chSequentialStreamWrite() instead of chprint() to send the message.

USB initialisation seems to need a one second sleep. Move it to
initialisation function. This slows down the whole init process but
then the USB subsystem is more ready to transmit messages on time.

There still seems to be some initial delay and if the main
application should wait a second or two before sending a burst of
messages over USB.