SDU: Remove extra delays in init.
MCU framework cannot spend time in initialisation while waiting for
USB to get ready. If main application sends messages when USB is not
ready, the buffers will overflow and data gets lost.

Also buffers seem to overflow if nobody is reading them on the other
end.

Therefore not asserting in the case of overflow.